100DaysofCode/Intermediate/Day-40-Capstone Part 2/Flight Club/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
twilio==8.10.2